backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
anthropic>=0.36.0

